requirements.txt
streamlit
opencv-python
numpy
scikit-image
matplotlib
mysql-connector-python
Pillow
#tambahan